requirements.txt
easyocr==1.7.2
editdistance==0.8.1
numpy==2.3.5
opencv_contrib_python==4.10.0.84
opencv_python==4.12.0.88
opencv_python_headless==4.12.0.88
paddleocr==3.2.0
paddlepaddle==3.2.0
Pillow==12.0.0
pytesseract==0.3.13
torch==2.2.2
transformers==4.57.3